Move some ui components to the screenshots module to serve as an example.
