+++ app/build.gradle
@@ -55,6 +55,8 @@ android {
 }
 
 dependencies {
+    implementation project(":screenshots")
+
     //Desugaring for use of java.time in api lower then 26
     coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:2.0.2'
