+++ screenshots/build.gradle
@@ -30,8 +30,6 @@ android {
 }
 
 dependencies {
-    implementation project(":app")
-
     def composeBom = platform('androidx.compose:compose-bom:2023.01.00')
     implementation composeBom
     implementation 'androidx.compose.material3:material3'
@@ -40,7 +38,6 @@ dependencies {
 
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'com.google.android.material:material:1.8.0'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
