+++ screenshots/src/main/java/com/appunite/loudius/screenshots/components/LoudiusOutlinedButton.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.appunite.loudius.ui.components
+package com.appunite.loudius.screenshots.components
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Icon
@@ -26,8 +26,8 @@ import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.appunite.loudius.R
-import com.appunite.loudius.ui.theme.LoudiusTheme
+import com.appunite.loudius.screenshots.R
+import com.appunite.loudius.screenshots.theme.LoudiusTheme
 
 enum class LoudiusOutlinedButtonStyle {
     Large,
