+++ screenshots/src/main/java/com/appunite/loudius/screenshots/components/LoudiusText.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.appunite.loudius.ui.components
+package com.appunite.loudius.screenshots.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.MaterialTheme
@@ -23,7 +23,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import com.appunite.loudius.ui.theme.LoudiusTheme
+import com.appunite.loudius.screenshots.theme.LoudiusTheme
 
 enum class LoudiusTextStyle {
     ListHeader,
