+++ screenshots/src/main/res/drawable/ic_github.xml
